Body: Add story
